refactor: :truck: moved individual components in their own files
